Adds numbers made of many ciphers

+++ sample/src/java/pl/edu/agh/tkk17/sample/Evaluator.java
@@ -2,12 +2,12 @@ package pl.edu.agh.tkk17.sample;
 
 public class Evaluator
 {
-    public static int evaluate(Node tree)
+    public static float evaluate(Node tree)
     {
         return evaluateRpn(tree);
     }
 
-    private static int evaluateRpn(Node node)
+    private static float evaluateRpn(Node node)
     {
         RpnEvaluatorVisitor visitor = new RpnEvaluatorVisitor();
         node.accept(visitor);
